+++ app/src/main/java/com/example/moneysnap/data/entity/Income.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class Income(
     @PrimaryKey (autoGenerate = true)
     val _id: Int,
-    val amount: Double, // 수입 금액
+    val amount: Int, // 수입 금액
     val source: String, // 수입 출처 (예: 월급, 보너스, 중고거래)
     val date: String, // 수입 날짜
     val description: String? = null, // 수입 상세 설명
